+++ lib/communication/linux/netadapterlist.cpp
@@ -124,7 +124,7 @@ namespace hbm {
 			sa_family_t family;
 			char buf[INET6_ADDRSTRLEN];
 
-			tAdapters adapterMap;
+			Adapters adapterMap;
 
 			if (::getifaddrs(&interfaces) < 0) {
 				::syslog(LOG_ERR, "Error calling getifaddrs!");
@@ -156,7 +156,7 @@ namespace hbm {
 									struct sockaddr_in ipv4_address;
 									std::memcpy(&ipv4_address, interface->ifa_addr, sizeof(ipv4_address));
 
-									ipv4Address_t addressWithNetmask;
+									Ipv4Address addressWithNetmask;
 									if (inet_ntop(family, &ipv4_address.sin_addr, buf, sizeof(buf))) {
 										addressWithNetmask.address = buf;
 									}
@@ -176,7 +176,7 @@ namespace hbm {
 									struct sockaddr_in6 ipv6_address;
 									std::memcpy(&ipv6_address, interface->ifa_addr, sizeof(ipv6_address));
 
-									ipv6Address_t address;
+									Ipv6Address address;
 									if (inet_ntop(family, &ipv6_address.sin6_addr, buf, sizeof(buf))) {
 										address.address = buf;
 									}
@@ -217,19 +217,19 @@ namespace hbm {
 			::freeifaddrs(interfaces);
 		}
 
-		NetadapterList::tAdapters NetadapterList::get() const
+		NetadapterList::Adapters NetadapterList::get() const
 		{
 			std::lock_guard < std::mutex > lock(m_adaptersMtx);
 			return m_adapters;
 		}
 
-		NetadapterList::tAdapterArray NetadapterList::getArray() const
+		NetadapterList::AdapterArray NetadapterList::getArray() const
 		{
 			std::lock_guard < std::mutex > lock(m_adaptersMtx);
-			tAdapterArray result;
+			AdapterArray result;
 			result.reserve(m_adapters.size());
 
-			for(tAdapters::const_iterator iter = m_adapters.begin(); iter!=m_adapters.end(); ++iter) {
+			for(Adapters::const_iterator iter = m_adapters.begin(); iter!=m_adapters.end(); ++iter) {
 				result.push_back(iter->second);
 			}
 
@@ -240,7 +240,7 @@ namespace hbm {
 		{
 			std::lock_guard < std::mutex > lock(m_adaptersMtx);
 
-			for (tAdapters::const_iterator iter = m_adapters.begin(); iter != m_adapters.end(); ++iter) {
+			for (Adapters::const_iterator iter = m_adapters.begin(); iter != m_adapters.end(); ++iter) {
 				if (iter->second.getName().compare(adapterName) == 0) {
 					return iter->second;
 				}
@@ -254,12 +254,30 @@ namespace hbm {
 		{
 			std::lock_guard < std::mutex > lock(m_adaptersMtx);
 
-			tAdapters::const_iterator iter = m_adapters.find(interfaceIndex);
+			Adapters::const_iterator iter = m_adapters.find(interfaceIndex);
 			if(iter==m_adapters.end()) {
 				throw hbm::exception::exception("invalid interface");
 			}
 
 			return iter->second;
+		}
+
+		std::string NetadapterList::checkSubnet(communication::Ipv4Address& requestedAddress) const
+		{
+			std::string requestedSubnet = requestedAddress.getSubnet();
+
+			for (communication::NetadapterList::Adapters::const_iterator adapterIter=m_adapters.begin(); adapterIter!=m_adapters.end(); ++adapterIter ) {
+				const communication::Netadapter& adapter = adapterIter->second;
+				communication::AddressesWithNetmask addresses = adapter.getIpv4Addresses();
+				
+				for (communication::AddressesWithNetmask::const_iterator addressIter = addresses.begin(); addressIter!=addresses.end(); ++addressIter) {
+					const communication::Ipv4Address& address = *addressIter;
+					if (requestedSubnet==address.getSubnet()) {
+						return adapter.getName();
+					}
+				}
+			}
+			return "";
 		}
 
 		void NetadapterList::update()