+++ changelog.md
@@ -3,6 +3,15 @@
 ## v1.0.1
  - Renamed some netlink event types
  
-## v
+## v1.0.2
  - New method for retrieving subnet of ipv4 address/netmask
+ - New method checking whether subnet of ipv4 address/netmask is already occupied by another interface 
+ - Rename class ipv4Address_t => Ipv4Address
+ - Rename class ipv6Address_t => Ipv6Address
+ - Rename class addressesWithPrefix_t to AddressesWithPrefix
+ - Rename class addressesWithNetmask_t to AddressesWithNetmask
+ - Rename header communication/ipv4Address_t.h to communication/ipv4Address.h
+ - Rename header communication/ipv6Address_t.h to communication/ipv6Address.h
+
+